requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# AI and Machine Learning
google-generativeai>=0.3.0
openai>=1.0.0  # Optional: for future OpenAI integration
langchain>=0.1.0
langchain-community>=0.0.10

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10  # OCR capabilities

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11

# Data Processing and Analytics
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # Excel support
xlsxwriter>=3.1.0

# Database
sqlalchemy>=2.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web and APIs
requests>=2.31.0
httpx>=0.25.0

# Security and Validation
cryptography>=41.0.0
pydantic>=2.0.0
validators>=0.22.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging and Monitoring
loguru>=0.7.0

# Export and Reporting
reportlab>=4.0.0  # PDF generation
jinja2>=3.1.0  # Template engine

# Optional: Advanced Features
scikit-learn>=1.3.0  # ML capabilities
transformers>=4.30.0  # NLP models
torch>=2.0.0  # Deep learning
torchvision>=0.15.0

# Development and Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Environment Management
python-decouple>=3.8